Take the tenant CR out of the certificate manager and render utils
The certificate manager only needed to know whether the cluster is
multi-tenant, and the two tenant-scoped render helpers moved to their own
Enterprise-only file.
